apps/backend_api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
argcomplete==3.6.2
asgiref==3.8.1
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
datamodel-code-generator==0.30.1
Django==5.2
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
genson==1.3.0
gunicorn==23.0.0
h11==0.16.0
html2image==2.0.6
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflect==7.5.0
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
MarkupSafe==3.0.2
more-itertools==10.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
psycopg==3.2.7
psycopg2==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
pytest==8.3.5
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sentry-sdk==2.27.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
tomli==2.2.1
typeguard==4.4.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
websocket-client==1.8.0
whitenoise==6.9.0
